+++ Code/memory_map.h
@@ -31,9 +31,9 @@ void memory_write_oled_command(uint8_t com);
 
 void memory_write_oled_data(uint8_t com);
 
-void memory_write_sram_data(uint8_t address, uint8_t data);
+void memory_write_sram_data(uint16_t address, uint8_t data);
 
-uint8_t memory_read_sram_data(uint8_t address);
+uint8_t memory_read_sram_data(uint16_t address);
 
 
 #endif
